requirements.txt
fastapi
uvicorn
python-dotenv
huggingface_hub
pydantic
